+++ module/core/derive_tools/changelog.md
@@ -81,3 +81,5 @@
 *   fix(derive_tools): Re-enable and fix all tests
 
 *   fix(derive_tools): Re-enable and fix all manual tests
+
+*   fix(derive_tools): Re-enable and fix basic tests